feat(mlcache) use an FFI struct to encode tables
This way we can avoid using a separate set of keys in the shm to store
if some values are encoded or not. If such a key would come to be purged
by the shm LRU mechanism, while the actual value is not, we would return
non-decoded table values (aka, JSON encoded).