feat(mlcache) isolated mlcache instances
Now, several instances of mlcache can all rely on the same underlying
shm, and maintain their own LRUs. There will be no conflicting keys in
the shm level because each mlcache namespaces its own keys inserted in
the shm.

A nice use-case for this is having several mlcache to hold different
types of data. Since the unit for the size of lua-resty-lru is the
number of slots, it is preferable to store objects of roughly the same
size, in order to be able to provide an estimate for the LRU size.

Ex:
    local cache_1 = mlcache.new("db_cache_shm", { lru_size = 100 })
    local cache_2 = mlcache.new("db_cache_shm", { lru_size = 10e5 })

Here, cache_1 could contain a few number of large objects (to not
exhaust the Lua VM memory), and cache_2 could contain a large number of
small objects without causing too much problems on memory pressure.