fix(get) cache for indefinite uses 'nil' in LRU instead of 'math.huge'
`nil` is the value lua-resty-lrucache uses to define an indefinite TTL. It
saves CPU cycles at retrieval because lua-resty-lrucache will compare
`ngx.now()` with the TTL if a TTL is set.

More importantly, this fixes an issue with the shm -> LRU setting when
computing the remaining TTL. When the TTL is `0` (indefinite as per this
library's convention), we used to set the value in the LRU with a
negative TTL (`0 - (now() - at)`), which resulted in the value actually
not being cached in the LRU because the TTL would be negative (already
expired).