fix(mlcache) do not error out on 'no memory'
When the shm is full and eviction must happen through the LRU mechanism,
it could be that the removed values do not free a large enough amount of
memory in the shm, thus leading `set()` to fail and return a "no memory"
error.

This can happen when values being stored have sizes of different orders of
magnitude.

We now ignore such errors, so that the user can still retrieve the data
and benefit from it in the L1 cache which does not suffer from such a
limitation.

We print a warning notice when such errors occur, but a new `opts.quiet`
option can disable this behavior.

From #41