fix(set) ensure 'no memory' errors are returned
When shm insertion fails inside of a `:get()` operation, we purposefully
catch 'no memory' errors and serve the data fetched from the callback
to ensure resiliency in the host application.

However, callers of `:set()` must be informed that the insertion
operation failed, and not set the value in the LRU if shm insertion
failed.