feat(peek) [breaking] use shm:ttl() in recent OpenResty versions
Use the new `shm:ttl()` API which was contributed in order to support
mlcache's `peek()` method. The existing `peek()` implementation was a
temporary solution until then. It is slightly unfortunate that we chose
to return the value along with the remaining TTL, as this behavior now
must break.

Pros/cons
---------
+ As a result, `peek()` is ligher and thus, faster
+ It does not touch the shm LRU queue
- It is a breaking change

Fix #39