Fix expiration when the remaining TTL is exactly 0
As 0 is used as a sentinel for "no TTL" when caching in the worker cache,
if the remaining TTL of a shared hit happens to be exactly 0* that entry
might be cached forever.

This commit fixes that behaviour by considering such cases the same way it
handles a stale value.

* this can happen as the expiration in the shared dict doesn't appear to
  be fully accurate