fix(purge) prevent purging when using a custom LRU cache
As it stands, calling `purge()` on an instance that has a custom LRU
cache will also garbage collect said instance, and a new one created by
`lrucache.new()` will take its place.

As long as the new `flush_all()` method for LRU caches isn't supported
(introduced in OpenResty 1.13.6.2), we shall be proactive about this
limitation and prevent the use of `purge()`.

This limitation is soon to be lifted for OpenResty >= 1.13.6.2 thanks
to #78.