feat(get) callback can now be 'nil' to probe L1/L2 caches
Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>

Specifying a `nil` callback to `:get()` is now possible. Doing so will
return any item from L1/L2 (and do the L2 -> L1 promotion as usual).

If a requested key is not in the cache, the return values will be `nil,
nil, -1`, where -1 is a new meaning for the `hit_lvl` return value
indicating a cache miss. This is helpful when trying to distinguish a
*negative hit* (cached `nil`) from a regular *miss* (no value). The
former will have a positive `hit_lvl` indicating the negative hit was
cached in L1/L2).

Example:

    local val, err, hit_lvl = cache:get("key")
    if val == nil then
        if hit_lvl == -1 then
            -- miss (no value)
        end

        -- negative hit (cached `nil`)
    end

Thanks Hamish Forbes and Corina Purcarea for the feature proposal and
patch.

From #96