feat(ipc) avoid new workers attempting to replay evicted events
A follow-up to #88.

Described in #93 and reported a few months ago is the likely possibility
that mlcache and ipc instances are created in the `init` phase. If a
worker is to be started much later in the master process' lifetime, the
newly forked ipc instance will have an `idx` attribute set to the
shm_idx value at the time of `init` (likely 0). These workers will
resume polling evicted events, and `poll()` will likely timeout
indefinitely from there.

For the fix in #88 to work, the mlcache instance has to be instantiated
during `init_worker` or later.

This patch proposes an approach which works for instances created in
both `init` and `init_worker`: as soon as the ipc shm has started
evicting items, we guarantee that future workers will resume polling at
the current index, without having to call any method but `poll()`.

Fix #93